gs, gcs: Move router_config generation to pfconfig
